Allow slides to be duplicated every N slides
This allows slides in any source to be marked to be duplicated
every N slides. So if the "duplicate" argument is set for a source,
all slides generated by that source will get marked to have a copy within
a block of "duplicate" slides. So if a slide has "duplicate" of 5, it will
have one copy in slides 0 to 4, one copy in slides 5 to 9, one in 10 to 14
and so on. The exact position is randomised within the block, to avoid
confusion caused by predictable ordering.